requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
django-cors-headers
django-dotenv
django-redis
psycopg2-binary
djangorestframework-simplejwt
PyJWT
drf-yasg
Celery
